requirements.txt
streamlit>=1.28.0
python-binance>=1.0.19
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
python-dotenv>=1.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.4.0
